+++ src/sqlite_anyio/__init__.py
@@ -6,7 +6,4 @@ from .sqlite import connect as connect
 from .sqlite import exception_logger as exception_logger
 
 
-try:
-    __version__ = importlib.metadata.version("sqlite_anyio")
-except importlib.metadata.PackageNotFoundError:  # pragma: nocover
-    __version__ = "unknown"
+__version__ = importlib.metadata.version("sqlite_anyio")