+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Version history
 
+## 0.2.8
+
+- Return self instead of new cursor.
+
+## 0.2.7
+
+- Add `Cursor` async context manager that closes the cursor on exit.
+
+## 0.2.6
+
+- Use `Self` in `Connection` async context manager.
+
 ## 0.2.5
 
 - Add trusted published, `CHANGELOG.md`.